+++ lib/firebase_auth_simplify.dart
@@ -7,4 +7,5 @@ export 'src/api/firebase_email_auth_api.dart';
 export 'src/api/firebase_google_auth_api.dart';
 export 'src/api/firebase_kakao_auth_api.dart';
 export 'src/api/firebase_facebook_auth_api.dart';
+export 'src/api/firebase_phone_auth_api.dart';
 export 'package:firebase_auth/firebase_auth.dart';